+++ geokey/projects/models.py
@@ -246,7 +246,7 @@ class Project(models.Model):
             not user.is_anonymous() and (
                 self.usergroups.filter(users=user).exists()))
 
-    def get_all_contributions(self, user):
+    def get_all_contributions(self, user, search=None):
         """
         Returns all contributions a user can access in a project. It gets
         the SQL clauses of all data groupings in the project and combines them
@@ -263,16 +263,14 @@ class Project(models.Model):
             List of geokey.contributions.models.Observations
         """
         data = None
-        if self.is_admin(user):
-            # Return everything for admins
-            return self.observations.for_moderator(user)
-        elif self.can_moderate(user):
+        is_admin = self.is_admin(user)
+        if is_admin or self.can_moderate(user):
             data = self.observations.for_moderator(user)
         else:
             data = self.observations.for_viewer(user)
 
         where_clause = None
-        if self.isprivate and not user.is_anonymous():
+        if not is_admin and self.isprivate and not user.is_anonymous():
             clauses = []
 
             for group in self.usergroups.filter(users=user):
@@ -282,10 +280,13 @@ class Project(models.Model):
             if clauses:
                 where_clause = '(' + ') OR ('.join(clauses) + ')'
 
-        if not where_clause:
-            return data
+        if where_clause:
+            data = data.extra(where=[where_clause]).distinct()
 
-        return data.extra(where=[where_clause]).distinct()
+        if search:
+            data = data.search(search)
+
+        return data
 
     def contact_admins(self, sender, mail_content):
         """